Work on import resolution, extern func calls, and string support
Implement rough draft of import resolution that loads a module of
functions into the AST immediately after the parsing stage and before
the type checking / inference stage. Add hardcoded external function
declarations to LLVM IR generation stage that allows for a dynamically
linked library compiled from C to be linked via clang at compile time.
Begin adding string support to all compiler stages although its current
state is not functional. Refactor project structure to contain a
frontend and backend for the compiler. The goal is to share the frontend
functionality between the current LLVM IR backend and to add an
interpreted version of the language that runs on a custom bytecode VM.